fix permissions on dashboard,add text watcher on cash payment

+++ app/src/main/java/com/pos/leaders/leaderspossystem/AddUserActivity.java
@@ -1,12 +1,10 @@
 package com.pos.leaders.leaderspossystem;
 
 import android.content.Intent;
-import android.graphics.drawable.ColorDrawable;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
-import android.view.ViewGroup;
 import android.view.Window;
 import android.view.WindowManager;
 import android.widget.AbsListView;
@@ -15,23 +13,16 @@ import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ListView;
-import android.widget.TextView;
 import android.widget.Toast;
 
-import com.pos.leaders.leaderspossystem.DataBaseAdapter.DepartmentDBAdapter;
 import com.pos.leaders.leaderspossystem.DataBaseAdapter.PermissionsDBAdapter;
 import com.pos.leaders.leaderspossystem.DataBaseAdapter.UserDBAdapter;
 import com.pos.leaders.leaderspossystem.DataBaseAdapter.UserPermissionsDBAdapter;
-import com.pos.leaders.leaderspossystem.Models.Department;
-import com.pos.leaders.leaderspossystem.Models.Permissions;
+import com.pos.leaders.leaderspossystem.Models.Permission.Permissions;
 import com.pos.leaders.leaderspossystem.Models.User;
-import com.pos.leaders.leaderspossystem.Tools.PermissionsGridViewAdapter;
-import com.pos.leaders.leaderspossystem.Tools.SESSION;
 import com.pos.leaders.leaderspossystem.Tools.TitleBar;
 
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Calendar;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -51,7 +42,7 @@ public class AddUserActivity extends AppCompatActivity {
 List<String> selectedFromList=new  ArrayList<String>();;
 	ArrayList<Integer> permissions_name;
 	List<String> permissionName;
-	Map<String,Integer> permissionsMap=new HashMap<String,Integer>();
+	Map<String,Long> permissionsMap=new HashMap<String,Long>();
 	ArrayAdapter<String> LAdapter;
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
